+++ examples/linkaudio/AudioPlatform_Portaudio.cpp
@@ -65,8 +65,8 @@ int AudioPlatform::audioCallback(const void* /*inputBuffer*/,
 
   for (unsigned long i = 0; i < inNumFrames; ++i)
   {
-    buffer[i * 2] = engine.mBuffer[i];
-    buffer[i * 2 + 1] = engine.mBuffer[i];
+    buffer[i * 2] = static_cast<float>(engine.mBuffer[i]);
+    buffer[i * 2 + 1] = static_cast<float>(engine.mBuffer[i]);
   }
 
   return paContinue;
